+++ lib/src/util/extensions/ext.dart
@@ -1,2 +1,3 @@
 export 'package:cubes/src/util/extensions/build_context_extensions.dart';
 export 'package:cubes/src/util/extensions/observable_value_extensions.dart';
+export 'package:cubes/src/util/extensions/string_extensions.dart';